+++ docs/source/user_guide/index.md
@@ -17,7 +17,7 @@ supported_systems
 tensor_types
 scalar_tensors
 matrix_vector
-flexible_tensors
+tensor
 compound_types
 static
 sub_functions
